telegram_bot/requirements.txt
aiofiles~=23.2.1
aiogram==3.12.0
aiogram-media-group==0.5.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
bitrix24-rest==1.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.4
psycopg2-binary==2.9.9
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
shortuuid==1.0.11
tortoise-orm==0.20.0
typing_extensions==4.7.1
urllib3==2.2.0
yarl==1.9.3
tbank_kassa_api==0.0.2.7.6
loguru